Set up app with RSpec/Capybara acceptance testing
Configure Capybara's app host to the development server (ran with yarn
start) to run the acceptance tests against.

As we're currently  using localStorage to persist itineraries, each
time the test are ran, a new chromedriver browser instance is initiated
which will have an empty localStorage. However, we'll need to clear the
localStorage manually between tests to ensure that is the case -
especially when test are ran randomly.